DALL_E_Image_Generation_WebApp/requirements.txt
openai==0.8.0
Pillow==9.3.0
streamlit==1.16.0
urllib3==1.26.12